WIP: handle alignement during evacuation
Since evacuation can relocate object to a start pointer with different
alignement, we need to recompute alignment and padding instead of
blindly copying the original block. This commit dumbly stores the
alignement as a few bits in the size in the block header.

However, it turned out to be insufficient: in fact, we need to
reconstruct the full layout of the original type. I think in the end I
can have only one true `alloc` implementation, taking an untyped u8 ptr
and a layout, with `alloc<T>` being just a wrapper for `alloc_layout(ptr
as *u8, Layout::new::<T>())`. Moving can't be just basic copy, it's
allocation as well.