First steps of mature space allocator
To properly handle the mature space, we need to have a proper allocator,
able to allocate out of free blocks, and to free allocated blocks. We're
going for the simplest design first, a basic linked-list allocator with
a first-fit strategy. This commit reworks the interface of the block
header and the spaces to make this possible, but the allocator is not
yet fully implemented.